test: add tests for `_prepareValidateOpenPosition` data (#403)
* test: add tests for the _prepareValidateOpenPositionData function

* fix(leverage): change _getLeverage to return a uint256 to avoid overflows

* test(leverage): add a check on the worst case scenario of _getLeverage

* refactor: remove unused Math library

* docs: fix NatSpec

+++ src/interfaces/UsdnProtocol/IUsdnProtocolTypes.sol
@@ -318,7 +318,7 @@ interface IUsdnProtocolTypes {
         bytes32 tickHash;
         Position pos;
         uint128 liqPriceWithoutPenalty;
-        uint128 leverage;
+        uint256 leverage;
         uint8 liquidationPenalty;
         bool isLiquidationPending;
     }
